@@ -1,9 +1,9 @@
 import 'package:flutter_svg/svg.dart';
-import 'package:js_trions/core/AppTheme.dart';
+import 'package:js_trions/core/app_theme.dart';
 import 'package:js_trions/ui/screens/AboutScreen.dart';
-import 'package:js_trions/ui/screens/DashboardScreen.dart';
+import 'package:js_trions/ui/screens/dashboard_screen.dart';
 import 'package:js_trions/ui/screens/ProjectsScreen.dart';
-import 'package:js_trions/ui/screens/SettingsScreen.dart';
+import 'package:js_trions/ui/screens/settings_screen.dart';
 import 'package:tch_appliable_core/tch_appliable_core.dart';
 import 'package:tch_common_widgets/tch_common_widgets.dart';
 
