@@ -1,5 +1,5 @@
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:js_trions/core/AppTheme.dart';
+import 'package:js_trions/core/app_theme.dart';
 import 'package:js_trions/model/Project.dart';
 import 'package:js_trions/ui/widgets/ChipWidget.dart';
 import 'package:tch_appliable_core/tch_appliable_core.dart';
@@ -132,6 +132,8 @@ class _ChipWidget extends StatelessWidget {
       case TranslationsJsonFormat.ObjectInside:
         text = tt('edit_project.field.translations_json_format.object_inside');
         break;
+      default:
+        debugPrint('TCH_e Unknown TranslationsJsonFormat $translationsJsonFormat');
     }
 
     return ChipWidget(
